[xla:runtime] Do not duplicate single-use argument loads out of the entry block
This is effectively deoptimizing hoisted loads out of loops.
Removing it allows us to achieve with XLA Runtime similar performance
as XLA:CPU Classic in math microbenchmarks.

This was originally introduced to reduce the size of the entry
block, which was causing overly long compilation times. It is likely
that this problem has now been fixed elsewhere, making the deoptimization
unnecessary.

PiperOrigin-RevId: 507495994